# taskgraph: Speed up gradle calls