Per https://github.com/mozilla-mobile/fenix/pull/5488#discussion_r327532557. The patch basically divides the duration of `taskgraph optimized --json --parameters main-repo-cron-nightly.yml` by 6: 

| Run | Current master                                | After this patch                             |
| --- | --------------------------------------------- | ---------------------------------------------| 
|  0  | 7.21s user 0.98s system 13% cpu 59.751 total  | 1.38s user 0.21s system 16% cpu 9.627 total  |
|  1  | 6.96s user 0.94s system 13% cpu 1:00.63 total | 1.52s user 0.23s system 18% cpu 9.626 total  |
|  2  | 7.19s user 0.96s system 13% cpu 1:01.62 total | 1.35s user 0.20s system 16% cpu 9.502 total  |
|  3  | 6.69s user 0.90s system 13% cpu 57.952 total  | 1.36s user 0.20s system 16% cpu 9.465 total  |
|  4  | 6.96s user 0.93s system 12% cpu 1:01.50 total | 1.41s user 0.21s system 17% cpu 9.107 total  |
|  5  | 7.41s user 0.99s system 13% cpu 1:02.76 total | 1.36s user 0.20s system 16% cpu 9.214 total  |
|  6  | 7.10s user 0.96s system 13% cpu 59.311 total  | 1.58s user 0.24s system 18% cpu 9.740 total  |
|  7  | 6.95s user 0.92s system 12% cpu 1:00.72 total | 1.35s user 0.20s system 16% cpu 9.242 total  |
|  8  | 6.89s user 0.92s system 12% cpu 1:00.63 total | 1.36s user 0.20s system 09% cpu 16.985 total |
|  9  | 6.79s user 0.91s system 13% cpu 59.158 total  | 1.73s user 0.26s system 19% cpu 10.220 total |
|  10 | 7.01s user 0.95s system 13% cpu 1:00.57 total | 1.36s user 0.21s system 15% cpu 10.075 total |

### Pull Request checklist
<!-- Before submitting the PR, please address each item -->
- [ ] **Quality**: This PR builds and passes detekt/ktlint checks (A pre-push hook is recommended)
- [ ] **Tests**: This PR includes thorough tests or an explanation of why it does not
- [ ] **Screenshots**: This PR includes screenshots or GIFs of the changes made or an explanation of why it does not
- [ ] **Accessibility**: The code in this PR follows [accessibility best practices](https://github.com/mozilla-mobile/shared-docs/blob/master/android/accessibility_guide.md) or does not include any user facing features

### After merge
- [ ] **Milestone**: Make sure issues finished by this pull request are added to the [milestone](https://github.com/mozilla-mobile/fenix/milestones) of the version currently in development.

### To download an APK when reviewing a PR:
1. click on Show All Checks,
2. click Details next to "Taskcluster (pull_request)" after it appears and then finishes with a green checkmark,
3. click on the "Fenix - assemble" task, then click "Run Artifacts".
4. the APK links should be on the left side of the screen, named for each CPU architecture